Proxy textures working
Finally figured out the problem I was having understanding reduz proxy function from 3.2, the parameters were misnamed.
Now it works, so single-window and also the 2d window work.

Only one bug now, they are upside down. This was expected as there was a Y flip flag that had changed between 3.2 and 4.0, and I left it as a todo in the source.

It will be close to a comment like this, if you search for them:

// FTODO
(four todo)